Start to extract the audio player into its own component
